requirements.txt
anyio==3.6.2
asgiref==3.6.0
astropy==5.2.1
autopep8==2.0.1
certifi==2022.12.7
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
jplephem==2.18
numpy==1.24.2
packaging==23.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyerfa==2.0.0.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
rfc3986==1.5.0
sgp4==2.21
six==1.16.0
skyfield==1.45
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
whitenoise==6.4.0
